Decount Ad
Decount has made a claim that Jiahua Chen was an employee when he created  this design and as a result had no right to release it under an MIT license. As I have gotten tired of hearing them argue with Jiahua, I have generously offered to put this little message in exchange for their silence. This in no way means that I believe their claims, but rather don't have any interest in continuing the debate.